+++ common/src/lib.rs
@@ -5,5 +5,4 @@ pub mod utils;
 
 pub mod config {
     pub mod types;
-    pub mod utils;
 }
