+++ common/src/utils.rs
@@ -13,3 +13,19 @@ pub fn address_to_hex_string(address: &Address) -> String {
 pub fn u64_to_hex_string(val: u64) -> String {
     format!("0x{val:x}")
 }
+
+pub fn bytes_deserialize<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+where
+    D: serde::Deserializer<'de>,
+{
+    let bytes: String = serde::Deserialize::deserialize(deserializer)?;
+    Ok(hex_str_to_bytes(&bytes).unwrap())
+}
+
+pub fn bytes_serialize<S>(bytes: &Vec<u8>, serializer: S) -> Result<S::Ok, S::Error>
+where
+    S: serde::Serializer,
+{
+    let bytes_string = hex::encode(bytes);
+    serializer.serialize_str(&bytes_string)
+}
