+++ config/src/lib.rs
@@ -17,3 +17,10 @@ pub use cli::*;
 /// Network Configuration
 pub mod networks;
 pub use networks::*;
+
+/// Generic Config Types
+pub mod types;
+pub use types::*;
+
+/// Generic Utilities
+pub mod utils;
